C API for signaling conditions
Because R's C API does not have a direct way to signal a condition, this
implementation simply constructs a condition object and calls the R
`signalCondition` function.

This file is located in `inst/include` so that other packages can use it
by adding `LinkingTo: rlang` to their DESCRIPTION. The function is
static so it will be copied to each compilation unit, avoiding multiple
definition errors. `inline` is not part of the c89 standard, so cannot
be used without `-std=c99` or compiler extensions.