Do not cross write barrier
Overscope rechaining used to circumvent the write barrier to avoid
generational adjustments, which I thought would be fine because we
were resetting to the original parent in a on.exit() call. On
hindsight, this was a bit reckless since I have very little idea about
how the GC works, and it doesn't make any difference in simple
benchmarks anyway.

cc @hadley