Fix handling of indent_status in deparse()
The increase_indent() function was actually creating an invalid
object via the expression "new_node(FALSE, 0L)".

The new_node function is a wrapper for the cons() function, which
expects two arguments:

  car: the value of the node to be created
  tail (or cdr): a pairlist (or NULL) which continues the list after the
                 new node created

The expression new_node(FALSE, 0L) then actually passes an integer vector
as the cdr or tail instead of a pairlist, creating an illegal object.

You can verify this by evaluating the following expressions:

     x <- new_node(FALSE, 0L)
     length(x[[1]])

Which will crash your R session.

The code does work at the moment because GNU R does not actually check
that the cdr argument to the cons() function is valid, and the value is
only accessed via the CDR macros, which also do not verify the expression.

Other interpreters (e.g. Renjin) are not so forgiving however.