Add `allow_na` to `check_logical()`
I think this finishes everything for #1724. The rest was fixed in #1742. A separate PR, #1745, additionally adds `allow_empty`, but I wanted to leave that out of this to make it easier to review the one small, relevant change.

I didn't add add anything to NEWS.md, since the standalone kinda has its own news. I'm not sure if it would be better to also mention is in the main NEWS.md to make it more visible.

Fixes #1724.